+++ did/did-agent-storage/lib/index.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-useless-constructor */
 /* eslint-disable no-unused-vars */
 const { EventEmitter } = require('events');