+++ did/did-auth-storage-mongo/lib/index.d.ts
@@ -3,7 +3,7 @@
 import { EventEmitter } from 'events';
 /**
  * Defines the interface of DID-Auth Token Storage
- * Which is used during the DID-Auth process in a dApp
+ * Which is used to persist state during the DID-Auth process in a dApp
  *
  * @class AuthStorage
  * @see @arcblock/did-auth-storage-firebase